+++ edflow/explore.py
@@ -7,27 +7,31 @@ import streamlit as st
 from edflow.config import parse_unknown_args, update_config
 
 import numpy as np
-from edflow.util import walk, pp2mkdtable
+from edflow.util import walk, pp2mkdtable, retrieve, get_leaf_names
+from edflow.util.edexplore import (
+    isimage,
+    isflow,
+    istext,
+    display_flow,
+    display_flow_on_image,
+)
 from edflow import get_obj_from_str
-
-
-def isimage(obj):
-    return (
-        isinstance(obj, np.ndarray)
-        and len(obj.shape) == 3
-        and obj.shape[2] in [1, 3, 4]
-    )
-
-
-def isflow(obj):
-    return isinstance(obj, np.ndarray) and len(obj.shape) == 3 and obj.shape[2] in [2]
-
-
-def istext(obj):
-    return isinstance(obj, (int, float, str, np.integer, np.float))
+from edflow.data.dataset_mixin import DatasetMixin
 
 
 def display_default(obj):
+    """Find out how item could be displayed
+
+    Parameters
+    ----------
+    obj : Any
+        Item of example
+
+    Returns
+    -------
+    str
+        One of "Image", "Text", "Flow", "None"
+    """
     if isimage(obj):
         return "Image"
     elif istext(obj):
@@ -39,6 +43,15 @@ def display_default(obj):
 
 
 def display(key, obj):
+    """Display item in streamlit
+
+    Parameters
+    ----------
+    key : str
+        Subheader to be displayed
+    obj : Any
+        Item of example to be displayed
+    """
     st.subheader(key)
     sel = selector(key, obj)
     if sel == "Text":
@@ -48,23 +61,155 @@ def display(key, obj):
         st.image((obj + 1.0) / 2.0)
 
     elif sel == "Flow":
-        import flowiz as fz
-
-        img = fz.convert_from_flow(obj)
-        st.image(img)
+        display_flow(obj, key)
 
 
 def selector(key, obj):
+    """Show select box to choose display mode of obj in streamlit
+
+    Parameters
+    ----------
+    key : str
+        Key of item to be displayed
+    obj : Any
+        Item to be displayed
+
+    Returns
+    -------
+    str
+        Selected display method for item
+    """
     options = ["Auto", "Text", "Image", "Flow", "None"]
     idx = options.index(display_default(obj))
     select = st.selectbox("Display {} as".format(key), options, index=idx)
     return select
 
 
-def show_example(dset, idx):
+def custom_visualizations(ex, config):
+    """Displays custom visualizations in streamlit
+
+    The visualizations can be inserted to the config via their import path.
+    Everyone can implement a custom visualization for an example.
+
+    The visualization functions must accept the example and config as positional
+    arguments.
+
+
+    Examples
+    --------
+
+    Add visualizations to the text box with their import path. For example:
+
+    .. code-block:: python
+
+        edflow.util.edexplore.display_flow_on_image
+
+
+    A valid visualization function could look like for example:
+
+    .. code-block:: python
+
+        import streamlit as st
+        from edflow.util.edexplore import isimage, st_get_list_or_dict_item
+
+        def my_visualization(ex, config):
+            st.write("This is my visualization")
+
+            image1, image1_key = st_get_list_or_dict_item(ex, "image1", filter_fn=isimage)
+
+            st.image((image1 + 1) / 2)
+            
+            image2 = ex["image2"]
+            image3 = ex["image3"]
+
+            st.image((image2 + 1) / 2)
+            st.image((image3 + 1) / 2)
+
+
+    Visualizations can be displayed by default, if they are specified in the config.
+    An example for the configuration yaml file would be:
+
+    .. code-block::
+
+        edexplore:
+            visualizations:
+                custom:
+                    vis1:
+                        path: my_package.visualizations.my_visualization
+
+
+    Parameters
+    ----------
+    ex : dict
+        Example to be visualized
+    config : dict
+        Edexplore config
+    """
+    st.header("Custom visualizations")
+    default_visualizations = retrieve(
+        config, "edexplore/visualizations/custom", default=dict()
+    )
+    import_paths = [
+        vis.get("path", None)
+        for vis in default_visualizations.values()
+        if isinstance(vis.get("path", None), str)
+    ]
+    visualizatins_str = st.text_input(
+        "Visualization import paths: comma separated", value=",".join(import_paths)
+    )
+
+    if len(visualizatins_str) == 0:
+        st.text(custom_visualizations.__doc__)
+
+    for vis in [vis for vis in visualizatins_str.split(",") if vis != ""]:
+        try:
+            st.subheader(vis)
+            impl = get_obj_from_str(vis, reload=True)
+            impl(ex, config)
+        except Exception as error:
+            st.write(error)
+
+
+ADDITIONAL_VISUALIZATIONS = {
+    "custom": custom_visualizations,
+    "optical_flow_on_image": display_flow_on_image,
+}
+
+
+def show_example(dset, idx, config):
+    """Show example of dataset
+
+    Parameters
+    ----------
+    dset : DatasetMixin
+        Dataset to be shown
+    idx : int
+        Index to be shown
+    config : dict
+        Config used to show example
+    """
+
     ex = dset[idx]
+
+    # additional visualizations
+    default_additional_visualizations = retrieve(
+        config, "edexplore/visualizations", default=dict()
+    ).keys()
+    additional_visualizations = st.sidebar.multiselect(
+        "Additional visualizations",
+        list(ADDITIONAL_VISUALIZATIONS.keys()),
+        default=default_additional_visualizations,
+    )
+    if len(additional_visualizations) > 0:
+        st.header("Additional visualizations")
+        for key in additional_visualizations:
+            ADDITIONAL_VISUALIZATIONS[key](ex, config)
+
+    # dataset items
     st.header("Keys")
     walk(ex, display, pass_key=True)
+
+    # summaries
     st.header("Summary")
     summary = pp2mkdtable(ex, jupyter_style=True)
     # print markdown summary on console for easy copy and pasting in readme etc
@@ -79,6 +224,15 @@ def _get_state(config):
 
 
 def explore(config, disable_cache=False):
+    """Explore dataset specified in config
+
+    Parameters
+    ----------
+    config : dict
+        Edflow config dict used to explore dataset
+    disable_cache : bool, optional
+        Disable cache while loading dataset, by default False
+    """
     if not disable_cache:
         get_state = st.cache(persist=False, allow_output_mutation=True)(_get_state)
     else:
@@ -100,7 +254,7 @@ def explore(config, disable_cache=False):
             idx = np.random.choice(len(dset))
         st.sidebar.text("Index: {}".format(idx))
 
-    show_example(dset, idx)
+    show_example(dset, idx, config)
 
     st.header("config")
     cfg_string = pp2mkdtable(config, jupyter_style=True)